chore: rename plugin-input.ts
