+++ README.md
@@ -19,7 +19,7 @@
 2. Update [compute.yml](./.github/workflows/compute.yml) with your plugin's name and update the `id`.
 3. Update [context.ts](./src/types/context.ts) with the events that your plugin will fire on.
 4. Update [manifest.json](./manifest.json) with a proper description of your plugin.
-5. Update [plugin-inputs.ts](./src/types/plugin-inputs.ts) to match the `with:` settings in your org or repo level configuration.
+5. Update [plugin-input.ts](./src/types/plugin-input.ts) to match the `with:` settings in your org or repo level configuration.
 
 - Your plugin config should look similar to this:
